- moved some functions from retro to utility.
